chore: move updateOutOfSlot and feeHistory to handlers
